refactoring and adding a test file (empty)
