+++ dashboard/streamlit.py
@@ -8,6 +8,8 @@ import streamlit as st
 import altair as alt
 from pymssql import connect, Connection, exceptions, Cursor
 
+from queries import (get_archival_data, get_latest_metrics)
+
 COLOUR_PALETTE = ["#84b067", "#a7de83", "#4b633b", "#2c3b23"]
 
 AltairChart = alt.Chart
@@ -117,25 +119,25 @@ def filter_by_plant(selected_plants: list, plant_metrics: pd.DataFrame, archive_
     return [plant_metrics, archive_plants]
 
 
-def get_latest_metrics(cursor: Cursor) -> pd.DataFrame:
-    """Function gets the latest plant health metrics including: temperature, soil moisture levels
-    plant name, time of recording and last_watered, and extracts these to a dataframe."""
-    query = f"""
-        SELECT pm.temperature, pm.soil_moisture, latest_recording_info.latest_time,
-          p.plant_name, pm.plant_id, pm.last_watered
-        FROM {environ['SCHEMA_NAME']}.plant_metric pm
-        JOIN (SELECT plant_id,
-            MAX(recording_taken) as latest_time
-            FROM {environ['SCHEMA_NAME']}.plant_metric
-            GROUP BY plant_id) as latest_recording_info
-        ON pm.plant_id = latest_recording_info.plant_id
-            AND pm.recording_taken = latest_recording_info.latest_time
-        JOIN {environ['SCHEMA_NAME']}.plant as p ON pm.plant_id = p.plant_id;
-        """
-    cursor.execute(query)
-    result = cursor.fetchall()
+# def get_latest_metrics(cursor: Cursor) -> pd.DataFrame:
+#     """Function gets the latest plant health metrics including: temperature, soil moisture levels
+#     plant name, time of recording and last_watered, and extracts these to a dataframe."""
+#     query = f"""
+#         SELECT pm.temperature, pm.soil_moisture, latest_recording_info.latest_time,
+#           p.plant_name, pm.plant_id, pm.last_watered
+#         FROM {environ['SCHEMA_NAME']}.plant_metric pm
+#         JOIN (SELECT plant_id,
+#             MAX(recording_taken) as latest_time
+#             FROM {environ['SCHEMA_NAME']}.plant_metric
+#             GROUP BY plant_id) as latest_recording_info
+#         ON pm.plant_id = latest_recording_info.plant_id
+#             AND pm.recording_taken = latest_recording_info.latest_time
+#         JOIN {environ['SCHEMA_NAME']}.plant as p ON pm.plant_id = p.plant_id;
+#         """
+#     cursor.execute(query)
+#     result = cursor.fetchall()
 
-    return pd.DataFrame(result)
+#     return pd.DataFrame(result)
 
 
 def plot_live_temp(data: pd.DataFrame) -> AltairChart:
@@ -193,6 +195,7 @@ def plot_live_moisture(data: pd.DataFrame) -> AltairChart:
 
 def get_data_plant_table(data: dict) -> pd.DataFrame:
     """Display a table with plant IDs and names in Streamlit."""
+    print(type(data))
     df = pd.DataFrame({
         'plant_id': data['plant_id'],
         'plant_name': data['plant_name']
@@ -239,27 +242,27 @@ def plot_last_watered(data: pd.DataFrame) -> AltairChart:
     return chart
 
 
-def get_archival_data(cursor: Cursor) -> pd.DataFrame:
-    """Function gets archival data including averages of temperature, soil_moisture."""
-    query = f"""
-        SELECT 
-            ROUND(pa.avg_temperature, 2) AS avg_temperature, 
-            ROUND(pa.avg_soil_moisture, 2) as avg_soil_moisture, 
-            p.plant_name, 
-            p.plant_id
-        FROM {environ['SCHEMA_NAME']}.Plants_archive as pa
-         JOIN (SELECT plant_id,
-            MAX(last_recorded) as latest_time
-            FROM {environ['SCHEMA_NAME']}.Plants_archive
-            GROUP BY plant_id) as last_recording
-        ON pa.plant_id = last_recording.plant_id
-        JOIN {environ['SCHEMA_NAME']}.plant as p
-            ON pa.plant_id = p.plant_id;
-        """
-    cursor.execute(query)
-    result = cursor.fetchall()
+# def get_archival_data(cursor: Cursor) -> pd.DataFrame:
+#     """Function gets archival data including averages of temperature, soil_moisture."""
+#     query = f"""
+#         SELECT
+#             ROUND(pa.avg_temperature, 2) AS avg_temperature,
+#             ROUND(pa.avg_soil_moisture, 2) as avg_soil_moisture,
+#             p.plant_name,
+#             p.plant_id
+#         FROM {environ['SCHEMA_NAME']}.Plants_archive as pa
+#          JOIN (SELECT plant_id,
+#             MAX(last_recorded) as latest_time
+#             FROM {environ['SCHEMA_NAME']}.Plants_archive
+#             GROUP BY plant_id) as last_recording
+#         ON pa.plant_id = last_recording.plant_id
+#         JOIN {environ['SCHEMA_NAME']}.plant as p
+#             ON pa.plant_id = p.plant_id;
+#         """
+#     cursor.execute(query)
+#     result = cursor.fetchall()
 
-    return pd.DataFrame(result)
+#     return pd.DataFrame(result)
 
 
 def create_avg_temp_line(data: pd.DataFrame) -> alt.Chart:
