+++ dashboard/test_queries.py
@@ -1,0 +1,1 @@
+"""Test file for queries.py"""
